Scraper.py
This mini project is an Indeed Job Scraper built with Python, Selenium, and BeautifulSoup. It automates the process of collecting job listings from Indeed
, extracting details such as job title, company, location, salary, rating, posting date, and job URL. Optionally, it can also fetch the full job description for each listing.

The results are saved into a CSV file, making it easy for job seekers, recruiters, or analysts to filter and analyze job market trends. This project demonstrates web scraping, concurrent processing, and data handling with Pandas, making it a practical mini project for learning automation and data collection with Python.